// src/DS/types/Typography.ts
export enum TextSize {
  Small = "small",
  Medium = "medium",
  Large = "large",
  Base = "base",
  ExtraLarge = "xl",
  DoubleExtraLarge = "2xl",
}

export enum TextColor {
  Default = "default",
  Light = "light",
  Primary = "primary",
  Secondary = "secondary",
  Tertiary = "tertiary",
  Accent = "accent",
  Warning = "warning",
  Error = "error",
  Success = "success",
}

export enum FontWeight {
  Light = "light",
  Normal = "normal",
  Medium = "medium",
  SemiBold = "semibold",
  Bold = "bold",
}